+++ notify/impl_test.go
@@ -295,7 +295,7 @@ func TestOpsGenie(t *testing.T) {
 func TestEmailConfigNoAuthMechs(t *testing.T) {
 
 	email := &Email{
-		conf: &config.EmailConfig{}, tmpl: &template.Template{}, logger: log.NewNopLogger(),
+		conf: &config.EmailConfig{AuthUsername: "test"}, tmpl: &template.Template{}, logger: log.NewNopLogger(),
 	}
 	_, err := email.auth("")
 	require.Error(t, err)
@@ -304,8 +304,9 @@ func TestEmailConfigNoAuthMechs(t *testing.T) {
 
 func TestEmailConfigMissingAuthParam(t *testing.T) {
 
+	conf := &config.EmailConfig{AuthUsername: "test"}
 	email := &Email{
-		conf: &config.EmailConfig{}, tmpl: &template.Template{}, logger: log.NewNopLogger(),
+		conf: conf, tmpl: &template.Template{}, logger: log.NewNopLogger(),
 	}
 	_, err := email.auth("CRAM-MD5")
 	require.Error(t, err)
@@ -322,6 +323,15 @@ func TestEmailConfigMissingAuthParam(t *testing.T) {
 	_, err = email.auth("PLAIN LOGIN")
 	require.Error(t, err)
 	require.Equal(t, err.Error(), "missing password for PLAIN auth mechanism; missing password for LOGIN auth mechanism")
+}
+
+func TestEmailNoUsernameStillOk(t *testing.T) {
+	email := &Email{
+		conf: &config.EmailConfig{}, tmpl: &template.Template{}, logger: log.NewNopLogger(),
+	}
+	a, err := email.auth("CRAM-MD5")
+	require.NoError(t, err)
+	require.Nil(t, a)
 }
 
 func TestVictorOpsCustomFields(t *testing.T) {
@@ -371,3 +381,67 @@ func TestVictorOpsCustomFields(t *testing.T) {
 	// Verify that a custom field was added to the payload and templatized.
 	require.Equal(t, "message", m["Field_A"])
 }
+<<<<<<< HEAD
+=======
+
+func TestTruncate(t *testing.T) {
+	testCases := []struct {
+		in string
+		n  int
+
+		out   string
+		trunc bool
+	}{
+		{
+			in:    "",
+			n:     5,
+			out:   "",
+			trunc: false,
+		},
+		{
+			in:    "abcde",
+			n:     2,
+			out:   "ab",
+			trunc: true,
+		},
+		{
+			in:    "abcde",
+			n:     4,
+			out:   "a...",
+			trunc: true,
+		},
+		{
+			in:    "abcde",
+			n:     5,
+			out:   "abcde",
+			trunc: false,
+		},
+		{
+			in:    "abcdefgh",
+			n:     5,
+			out:   "ab...",
+			trunc: true,
+		},
+		{
+			in:    "a⌘cde",
+			n:     5,
+			out:   "a⌘cde",
+			trunc: false,
+		},
+		{
+			in:    "a⌘cdef",
+			n:     5,
+			out:   "a⌘...",
+			trunc: true,
+		},
+	}
+
+	for _, tc := range testCases {
+		t.Run(fmt.Sprintf("truncate(%s,%d)", tc.in, tc.n), func(t *testing.T) {
+			s, trunc := truncate(tc.in, tc.n)
+			require.Equal(t, tc.trunc, trunc)
+			require.Equal(t, tc.out, s)
+		})
+	}
+}
+>>>>>>> 0a2d108bb9b5cd1efd7d3d868cf8632ecf4cf6f8